+++ mtg_proxy_printer/model/carddb.sql
@@ -14,7 +14,7 @@
 -- along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 
-PRAGMA user_version = 0000033;
+PRAGMA user_version = 0000034;
 PRAGMA foreign_keys = on;
 PRAGMA journal_mode = 'wal';
 BEGIN TRANSACTION;
@@ -120,6 +120,8 @@ CREATE TABLE PrintingDisplayFilter (
   PRIMARY KEY (printing_id, filter_id)
 ) WITHOUT ROWID;
 
+CREATE INDEX PrintingDisplayFilter_Printing_from_filter_lookup ON PrintingDisplayFilter(filter_id);
+
 CREATE VIEW HiddenPrintingIDs AS
 SELECT printing_id
   FROM PrintingDisplayFilter
@@ -152,13 +154,22 @@ CREATE INDEX CardFace_for_translation ON CardFace(face_name_id, face_number, pri
 CREATE INDEX LookupPrintingBySet ON Printing(set_id);  -- Used by set code card filter logic
 
 CREATE VIEW VisiblePrintings AS
-WITH double_faced_printings(printing_id, is_dfc) AS (
+WITH
+  double_faced_printings(printing_id, is_dfc) AS (
   SELECT DISTINCT printing_id, TRUE as is_dfc
     FROM CardFace
-    WHERE is_front IS FALSE)
+    WHERE is_front IS FALSE),
+
+  token_printings(printing_id, is_token) AS (
+  SELECT printing_id, TRUE AS is_token
+    FROM DisplayFilters
+    JOIN PrintingDisplayFilter USING (filter_id)
+    WHERE filter_name = 'hide-token')
+
   SELECT card_name, set_code, set_name, "language", collector_number, scryfall_id, highres_image, face_number,
     is_front, is_oversized, png_image_uri, oracle_id, release_date, wackiness_score, release_date,
-    coalesce(double_faced_printings.is_dfc, FALSE) as is_dfc
+    coalesce(double_faced_printings.is_dfc, FALSE) as is_dfc,
+	coalesce(token_printings.is_token, FALSE) as is_token
   FROM Card
   JOIN Printing USING (card_id)
   JOIN MTGSet   USING (set_id)
@@ -166,18 +177,28 @@ WITH double_faced_printings(printing_id, is_dfc) AS (
   JOIN FaceName USING (face_name_id)
   JOIN PrintLanguage USING (language_id)
   LEFT OUTER JOIN double_faced_printings USING (printing_id)
+  LEFT OUTER JOIN token_printings USING (printing_id)
   WHERE Printing.is_hidden IS FALSE
     AND FaceName.is_hidden IS FALSE
 ;
 
 CREATE VIEW AllPrintings AS
-WITH double_faced_printings(printing_id, is_dfc) AS (
+WITH
+double_faced_printings(printing_id, is_dfc) AS (
   SELECT DISTINCT printing_id, TRUE as is_dfc
     FROM CardFace
-    WHERE is_front IS FALSE)
+    WHERE is_front IS FALSE),
+
+token_printings(printing_id, is_token) AS (
+  SELECT printing_id, TRUE AS is_token
+    FROM DisplayFilters
+    JOIN PrintingDisplayFilter USING (filter_id)
+    WHERE filter_name = 'hide-token')
+
   SELECT card_name, set_code, set_name, "language", collector_number, scryfall_id, highres_image, face_number,
      is_front, is_oversized, png_image_uri, oracle_id, release_date, wackiness_score, Printing.is_hidden,
-     coalesce(double_faced_printings.is_dfc, FALSE) as is_dfc
+     coalesce(double_faced_printings.is_dfc, FALSE) as is_dfc,
+	 coalesce(token_printings.is_token, FALSE) as is_token
   FROM Card
   JOIN Printing USING (card_id)
   JOIN MTGSet   USING (set_id)
@@ -185,6 +206,7 @@ WITH double_faced_printings(printing_id, is_dfc) AS (
   JOIN FaceName USING (face_name_id)
   JOIN PrintLanguage USING (language_id)
   LEFT OUTER JOIN double_faced_printings USING (printing_id)
+  LEFT OUTER JOIN token_printings USING (printing_id)
 ;
 
 CREATE VIEW CurrentlyEnabledSetCodeFilters AS