+++ mtg_proxy_printer/units_and_sizes.py
@@ -24,7 +24,10 @@ except ImportError:  # Compatibility with Python < 3.11
     from typing_extensions import NotRequired
 
 import pint
-from pint.facets.plain.registry import QuantityT, UnitT
+try:
+    from pint.facets.plain.registry import QuantityT, UnitT
+except ImportError:  # Compatibility with Pint 0.21 for Python 3.8 support
+    QuantityT = UnitT = typing.Any
 from PySide6.QtCore import QSize