+++ doc/changelog.md
@@ -4,21 +4,23 @@
 
 ## New features
 
-- Localization support. Translations are managed on [Crowdin.com](https://crowdin.com/project/mtgproxyprinter).   
-  Join, if you want to contribute translations :)
+- Localization support. Translations are managed on [Crowdin.com](https://crowdin.com/project/mtgproxyprinter). 
+    - Join there, if you want to contribute translations :)
     - Interface language is chosen based on the system locale, but can be explicitly set in the application settings.
     - Currently, the app includes translations into US English and German.
     - Goal is translations into all languages in which Magic sets get printed.
 - The deck import wizard can now directly download Scryfall search queries as deck lists
     - Added a text field to enter a Scryfall card search query, a button to show the result on the Scryfall website,
       and a button that downloads the search result as a deck list.
+    - Downloaded search results are treated as a list of singleton cards.
 - Add option to fully automatically remove basic lands from all imported deck lists.
     - When enabled in the settings, basic lands are automatically stripped from deck lists.
       Otherwise, the previous behavior, offering removal via a button click, is retained.
     - The option honors the settings regarding inclusion of Wastes or Snow-Covered basic lands.
 - Add new card filter to hide Art Series cards, which can be enabled in the application settings.
     - When updating from previous versions, the filter becomes functional after the next card data update.
-- Add a live-updating preview to the document settings window. 
+- Add a live-updating preview to the document settings window.
+- The app now has an icon. Provided by [islanders2013](https://www.reddit.com/user/islanders2013/)
 
 ## Changed features
 
@@ -35,6 +37,9 @@
 
 ## Fixed issues
 
+- Prefer cards to tokens with the same name when selecting a printing during deck list imports.
+  This prevents the app from choosing the token for cards that can create token copies of themselves.
+  This started to be an issue with the release of the Bloomborrow set.
 - Improved performance of the image cleanup wizard, if there are many images of non-English cards stored on disk.
 - Reworded and clarified some texts in the user interface, fixed grammar and spelling mistakes
 - The main window no longer stays open unresponsive for multiple seconds when trying to exit the application